+++ CHANGELOG.md
@@ -10,10 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - changelog
+- can list all existing users
 
 ### Fixed
 
 - 'bson' made mandatory, since it is used in all backends
+- sled backend fixed
 
 ### Changed